+++ reflectable/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Migrate the generated code and its dependencies to null safety.
 * Remove `bestEffortReflectedType`, `hasBestEffortReflectedType`,
   deprecated since version 0.5.0.
+* Change the return type of `ObjectMirror.delegate` from `dynamic` to
+  `Object?`, for consistency with several methods named `invoke...`.
 
 ## 2.2.9